Fix qiskit requirement
The JKU provider currently says it requires the 'qiskit' package,
however this is not correct. The 'qiskit' package is just a
meta-package for installation convenience that will install all the
qiskit elements. The JKU provider actually depends on qiskit-terra which
is the package for the terra code. Having this incorrect causes the
metapackage to be installed which forces specific versions of the
elements to be installed, see Qiskit/qiskit#222 for the issues this
can cause.