Added paragraph to README for retirement of qiskit-jku-provider (#75)
The qiskit-jku-provider is based on an out-dated implementation of the JKU simulator. The recent version is available as JKQ DDSIM.

I would suggest archiving this repository and adding a few lines of text of explanation with this PR.

JKQ DDSIM should be faster (although I haven't directly compared to the code in this repository) and it should be easier to maintain and extend. For example, a backend to construct the unitary is available.

Apart from not being actively maintained, the qiskit-jku-provider uses an out-dated implementation of the underlying decision diagrams as well as the simulation front-end. Additionally, the bindings between python and c++ are realized by writing to files from python, reading them in c++, and reading back the c++ output to python---the new implementation used pybind11.

* Added paragraph to README for retirement

* Added notes to migrate